requirements.txt
Django==4.2.2
djangorestframework==3.14.0
python-decouple==3.8
psycopg2-binary
drf-nested-routers
drf-yasg
django-cors-headers
gunicorn==20.1.0
django-filter==23.2
coreapi==2.3.3